complete moving, fix all clippy warnings, update enclave-runtime
