+++ enclave-runtime/src/lib.rs
@@ -98,8 +98,6 @@ mod beefy_merkle_tree;
 pub mod cert;
 pub mod error;
 pub mod rpc;
-mod sidechain_block_composer;
-mod sidechain_impl;
 mod sync;
 pub mod tls_ra;
 pub mod top_pool_execution;
