+++ sidechain/consensus/aura/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2018"
 [dependencies]
 log = { version = "0.4.14", default-features = false }
 finality-grandpa = { version = "0.14.3", default-features = false, features = ["derive-codec"] }
+sgx-externalities = { default-features = false, git = "https://github.com/integritee-network/sgx-runtime", branch = "master", optional = true }
 
 # sgx deps
 sgx_tstd = { branch = "v1.1.4-testing", git = "https://github.com/apache/teaclave-sgx-sdk.git", optional = true }
@@ -21,6 +22,8 @@ itp-settings = { path = "../../../core-primitives/settings", default-features = 
 itp-sgx-crypto = { path = "../../../core-primitives/sgx/crypto", default-features = false }
 itp-stf-state-handler = { path = "../../../core-primitives/stf-state-handler", default-features = false }
 itp-storage-verifier = { path =  "../../../core-primitives/storage-verified", default-features = false }
+itp-time-utils = { path = "../../../core-primitives/time-utils", default-features = false }
+itp-types = { path = "../../../core-primitives/types", default-features = false }
 its-block-composer = { path = "../../block-composer", default-features = false }
 its-consensus-common = { path = "../common", default-features = false }
 its-consensus-slots = { path = "../slots", default-features = false }
@@ -32,7 +35,6 @@ its-validateer-fetch = { path = "../../validateer-fetch", default-features = fal
 [dev-dependencies]
 itp-test = { path = "../../../core-primitives/test" }
 sp-keyring = { version = "4.0.0-dev", git = "https://github.com/paritytech/substrate.git", branch = "master"}
-itp-types = { path = "../../../core-primitives/types" }
 env_logger = "0.9.0"
 itp-storage = { path = "../../../core-primitives/storage" }
 
@@ -42,6 +44,8 @@ std = [
    "itp-sgx-crypto/std",
    "itp-stf-state-handler/std",
    "itp-storage-verifier/std",
+   "itp-time-utils/std",
+   "itp-types/std",
    "its-block-composer/std",
    "its-consensus-common/std",
    "its-consensus-slots/std",
@@ -50,13 +54,16 @@ std = [
    "its-top-pool-executor/std",
    "its-validateer-fetch/std",
    "log/std",
+   "sgx-externalities/std",
 ]
 sgx = [
    "sgx_tstd",
    "itp-sgx-crypto/sgx",
    "itp-stf-state-handler/sgx",
+   "itp-time-utils/sgx",
    "its-block-composer/sgx",
    "its-consensus-common/sgx",
    "its-consensus-slots/sgx",
    "its-top-pool-executor/sgx",
+   "sgx-externalities/sgx",
 ]
