+++ sidechain/consensus/aura/src/lib.rs
@@ -31,6 +31,7 @@ extern crate sgx_tstd as std;
 
 use core::marker::PhantomData;
 use itp_storage_verifier::GetStorageVerified;
+use itp_time_utils::duration_now;
 use its_consensus_common::{Environment, Error as ConsensusError, Proposer};
 use its_consensus_slots::{SimpleSlotWorker, Slot, SlotInfo};
 use its_primitives::{
@@ -180,7 +181,7 @@ where
 	}
 
 	fn proposing_remaining_duration(&self, slot_info: &SlotInfo<PB>) -> Duration {
-		proposing_remaining_duration(slot_info, its_consensus_slots::duration_now())
+		proposing_remaining_duration(slot_info, duration_now())
 	}
 }
 
@@ -244,7 +245,7 @@ mod tests {
 	use crate::mock::{default_header, validateer, EnvironmentMock, TestAura, SLOT_DURATION};
 	use itp_test::mock::onchain_mock::OnchainMock;
 	use itp_types::Block as ParentchainBlock;
-	use its_consensus_slots::{duration_now, PerShardSlotWorkerScheduler};
+	use its_consensus_slots::PerShardSlotWorkerScheduler;
 	use sp_keyring::ed25519::Keyring;
 
 	fn get_aura(onchain_mock: OnchainMock) -> TestAura {
